Added Student home UI
