+++ app/src/main/java/com/example/ucms_android/RegisterActivity.java
@@ -14,6 +14,7 @@ import com.example.ucms_android.model.ApiResponse;
 import com.example.ucms_android.model.RegisterRequest;
 import com.example.ucms_android.network.ApiClient;
 import com.example.ucms_android.network.AuthService;
+import com.example.ucms_android.ui.auth.LoginActivity;
 import com.google.android.material.button.MaterialButton;
 import com.google.android.material.snackbar.Snackbar;
 import com.google.android.material.textfield.TextInputEditText;
