+++ app/src/main/res/drawable/bg_stat_card.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <!-- 1. Soft Shadow Layer -->
+    <item
+        android:left="2dp"
+        android:top="2dp"
+        android:right="2dp"
+        android:bottom="4dp">
+        <shape android:shape="rectangle">
+            <solid android:color="#0D000000" />
+            <corners android:radius="@dimen/corner_radius_card" />
+        </shape>
+    </item>
+
+    <!-- 2. Orange Shape (Bottom Layer) -->
+    <item
+        android:right="2dp"
+        android:bottom="2dp">
+        <shape android:shape="rectangle">
+            <solid android:color="@color/colorPrimary" />
+            <corners android:radius="@dimen/corner_radius_card" />
+        </shape>
+    </item>
+
+    <!-- 3. White Shape (Top Layer) -->
+    <!-- This creates the crescent orange effect on the left -->
+    <item
+        android:left="12dp"
+        android:right="2dp"
+        android:bottom="2dp">
+        <shape android:shape="rectangle">
+            <solid android:color="@color/white" />
+            <corners android:radius="@dimen/corner_radius_card" />
+        </shape>
+    </item>
+
+</layer-list>
