+++ app/src/main/res/drawable/ic_notifications.xml
@@ -8,5 +8,5 @@
         android:strokeWidth="1.8"
         android:strokeLineCap="round"
         android:strokeLineJoin="round"
-        android:pathData="M12,5c-4.42,0 -8,0.9 -8,2s3.58,2 8,2 8,-0.9 8,-2 -3.58,-2 -8,-2z M4,7v10c0,1.1 3.58,2 8,2s8,-0.9 8,-2V7 M4,12c0,1.1 3.58,2 8,2s8,-0.9 8,-2"/>
+        android:pathData="M12,4c-3.3,0 -6,2.7 -6,6v7h12v-7c0,-3.3 -2.7,-6 -6,-6z M12,2v2 M4,17h16 M10,17c0,1.1 0.9,2 2,2s2,-0.9 2,-2" />
 </vector>
